refactor: 도서에 대한 독서 감상문 데이터 별도 컴포넌트로 분리 (#29)
