+++ src/components/composites/book/detail/record/list/BookRecordList.tsx
@@ -1,12 +1,9 @@
 import React from 'react';
 
-import { useUser } from '@/contexts';
-import { Dropdown, NoData, Profile } from '@/components';
-import { useCreateLikeForRecord } from '@/services';
-import LikeIcon from '@/assets/icon/ic_thumb_up.svg?react';
-import RatingIcon from '@/assets/icon/ic_rating.svg?react';
+import { Dropdown, NoData } from '@/components';
 import { BOOK_REVIEW_DROPDOWN_OPTIONS } from '@/constants';
 import type { GetBookUserRecordsServerModel, SelectOptionType } from '@/types';
+import BookRecordListItem from './item/BookRecordListItem';
 import * as S from './BookRecordList.styled';
 
 interface BookRecordListProps {
@@ -20,16 +17,6 @@ const BookRecordList = ({
   records,
   onSelect,
 }: BookRecordListProps) => {
-  const { user } = useUser();
-
-  const { mutate: createLikeRecord } = useCreateLikeForRecord();
-
-  const handleLikeButtonClick = (recordId: string) => () => {
-    const req = { recordId, userId: user?.id! };
-
-    createLikeRecord(req);
-  };
-
   return (
     <S.RecordSection>
       <S.RecordHeader>
@@ -43,33 +30,7 @@ const BookRecordList = ({
       </S.RecordHeader>
       {records.length ? (
         records.map((record) => (
-          <S.RecordList key={record.id}>
-            <S.RecordItemHeader>
-              <S.PersonInfo>
-                <Profile src={record.users.profile_url} />
-                <S.UserName>{record.users.nickname}</S.UserName>
-                <S.CreatedDate>{record.created_at.split('T')[0]}</S.CreatedDate>
-              </S.PersonInfo>
-              <S.RatingInfo>
-                {Array.from({ length: 5 }, (_, i) => (
-                  <RatingIcon
-                    key={i}
-                    css={S.ratingIcon(i < (record.rating ?? 0))}
-                  />
-                ))}
-              </S.RatingInfo>
-            </S.RecordItemHeader>
-            <S.RecordItemContent>{record.record_comment}</S.RecordItemContent>
-            <S.RecordItemFooter>
-              <S.LikeButton
-                type="button"
-                onClick={handleLikeButtonClick(record.id)}
-              >
-                <LikeIcon css={S.likeIcon} />
-                <S.Like>{record.like_count}</S.Like>
-              </S.LikeButton>
-            </S.RecordItemFooter>
-          </S.RecordList>
+          <BookRecordListItem key={record.id} record={record} />
         ))
       ) : (
         <NoData
